Rework build and update depedencies to enable Github Actions
- Updates muparserx dependency to 4.0.12 with Unicode and C++ 17 support
- Adds per-project build support with more relative pathing usage
- Removes the need for static VC++ dlls
- Adds qmake support for DeviceSelector and UpdateChecker
- Adds support for double processing in the Editor
- Updates Windows SDK APO libraries to newer versions
- Standardises on C++ 17/C17 builds
- Uses Unicode variants for all dependencies and EqualiserAPO
- Uses environmental variables for all projects for simpler Github Actions support
- Enables AVX2 for all components including the Editor

+++ Editor/guis/LoudnessCorrectionFilterGUI.cpp
@@ -22,7 +22,7 @@
 #include "LoudnessCorrectionFilterGUI.h"
 #include "ui_LoudnessCorrectionFilterGUI.h"
 
-LoudnessCorrectionFilterGUI::LoudnessCorrectionFilterGUI(float refLevel, float refOffset, float att)
+LoudnessCorrectionFilterGUI::LoudnessCorrectionFilterGUI(double refLevel, double refOffset, double att)
 	: IFilterGUI(),
 	ui(new Ui::LoudnessCorrectionFilterGUI)
 {
@@ -99,7 +99,7 @@ void LoudnessCorrectionFilterGUI::on_calibrateButton_clicked()
 
 void LoudnessCorrectionFilterGUI::updateVolume()
 {
-	float volume;
+	double volume;
 	HRESULT res = volumeController.getVolume(volume);
 
 	if (SUCCEEDED(res) && volume != lastVolume)
