+++ filters/IfFilterFactory.cpp
@@ -203,7 +203,7 @@ bool IfFilterFactory::toBoolean(const Value& value)
 	{
 	case 'i':
 		{
-			int i = value.GetInteger();
+			int i = (int)value.GetInteger();
 			result = (i != 0);
 		}
 		break;
